# doc/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Spika for Business API v3
  description: >-
    New set of API to use Spika for Business. First of all, you must get an
    access token by calling sign in API. 
schemes:
  - http
  - https
host: 'spika.chat'
basePath: /api/v3
...
paths:
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: Get list of all rooms
      description: >-
        Get list of all rooms of user pecified by access token. You can filter
        field by adding item to query. If you want to get all details of rooms,
        unnecessary to add query items.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: query
          name: keyword
          description: Search keyword to get room
          type: string
        - in: query
          name: offset
          description: The offset number of get history
          type: integer
          format: int32
        - in: query
          name: limit
          description: The limit number of get history
          type: integer
          format: int32
        - in: query
          name: sort
          description: >-
            Sort by fields set by string like [field name]:asc, [field
            name]:desc
          type: string
        - in: query
          name: fields
          description: >-
            Use for filtering params. Only return set fields if you set them,
            otherwise all fields will be returned.
          type: array
          items:
            type: string
      responses:
        '200':
          description: A list of Room
          schema:
            properties:
              rooms:
                type: array
                items:
                  $ref: '#/definitions/Room'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '500':
          description: Server error
      security:
        - api_key: []
    post:
      tags:
        - Rooms
      summary: Create room
      description: Create a new room belonging the user who have access token.
      consumes:
        - multipart/form-data
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: formData
          type: string
          name: name
          description: Name of new room, if you don't set a name, it will be [username]\'s New Room.
        - in: formData
          type: string
          name: description
          description: Descripiton of new room
        - in: formData
          type: string
          name: users
          description: User id list of new room by string separated by comma. The user who create room will be belong to this room even if you don't set any users.
        - in: formData
          type: file
          name: avatar
          description: Image file of avatar of new room
      responses:
        '200':
          description: Created new room
          schema:
            properties:
              room:
                $ref: '#/definitions/Room'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
  '/rooms/{roomId}':
    get:
      tags:
        - Rooms
      summary: Get room
      description: >-
        Get room object specified by room id. You can filter field by adding
        item to query. If you want to get all details of room, unnecessary to add query items.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room list
          type: string
          required: true
        - in: query
          name: fields
          description: >-
            Use for filtering params. Only return set fields if you set them,
            otherwise all fields will be returned.
          type: array
          items:
            type: string
      responses:
        '200':
          description: Document of room
          schema:
            properties:
              room:
                $ref: '#/definitions/Room'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
    put:
      tags:
        - Rooms
      summary: Update room details
      description: Update room details specified by id
      consumes:
        - multipart/form-data
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room
          type: string
          required: true
        - in: formData
          type: string
          name: name
          description: Name of new room, if you don't set a name, it will be [username]\'s New Room.
        - in: formData
          type: string
          name: description
          description: Descripiton of new room
        - in: formData
          type: file
          name: avatar
          description: Image file of avatar of new room
      responses:
        '200':
          description: Updated document of room
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
    delete:
      tags:
        - Rooms
      summary: Delete room
      description: Delete room corresonds to room id.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room
          type: string
          required: true
      responses:
        '200':
          description: Deleted room
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
  /rooms/leave:
    post:
      tags:
        - Rooms
      summary: Leave room
      description: Leave room. If you are the owner of room, the room and it's chat history will be deleted.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: body
          name: room id
          description: Parameters of the new room.
          schema:
            properties:
              roomId:
                type: string
          required: true
      responses:
        '200':
          description: Leave room
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
  '/rooms/{roomId}/users/':
    get:
      tags:
        - Rooms
      summary: Get user list of room
      description: Get user list of room specified by id. Only return user ids.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room
          type: string
          required: true
      responses:
        '200':
          description: User list of room
          schema:
            properties:
              room:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  users:
                    $ref: '#/definitions/UserList'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
    post:
      tags:
        - Rooms
      summary: Invite user to room
      description: Add user to room specified by id
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room list
          type: string
          required: true
        - in: body
          name: body
          schema:
            properties:
              users:
                $ref: '#definitions/UserList'
      responses:
        '200':
          description: Added users to room
          schema:
            properties:
              room:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  users:
                    $ref: '#/definitions/UserList'
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
    delete:
      tags:
        - Rooms
      summary: Kick user out from room
      description: >-
        Delete user from room specified by id. It actually update details of
        gorup.
      parameters:
        - in: header
          name: access-token
          description: Access token which is generated in signin.
          type: string
          required: true
        - in: path
          name: roomId
          description: room ID to get room list
          type: string
          required: true
        - in: body
          name: body
          schema:
            $ref: '#definitions/UserList'
      responses:
        '200':
          description: Deleted users
        '401':
          description: Wrong apikey
        '403':
          description: Wrong access token
        '422':
          description: Bad parmeter
        '500':
          description: Server error
      security:
        - api_key: []
...
securityDefinitions:
  api_key:
    type: apiKey
    name: apiKey
    in: header
definitions:
...
  UserList:
    type: array
    items:
      type: string
...
  Room:
    properties:
      id:
        type: string
      users:
        $ref: '#/definitions/UserList'
      owner:
        type: object
        properties:
          type:
            type: string
          index:
            type: boolean
            example: true
      name:
        type: string
      description:
        type: string
      created:
        type: number
        format: double
      modified:
        type: number
        format: double
      lastMessage:
        type: string
        description: id of last message
      type:
        type: number
        format: double
      avatar:
        $ref: '#/definitions/Avatar'
...
  Avatar:
    properties:
      picture:
        type: object
        properties:
          originalName:
            type: string
          size:
            type: integer
            format: int32
          mimeType:
            type: string
          nameOnServer:
            type: string
      thumbnail:
        type: object
        properties:
          originalName:
            type: string
          size:
            type: integer
            format: int32
          mimeType:
            type: string
          nameOnServer:
            type: string
...